+++ server/libs/platform/platform-component/platform-component-registry/platform-component-registry-service/src/main/java/com/bytechef/platform/component/registry/handler/loader/AbstractComponentHandlerLoader.java
@@ -21,8 +21,8 @@ import com.bytechef.commons.util.OptionalUtils;
 import com.bytechef.component.ComponentHandler;
 import com.bytechef.component.definition.ActionDefinition;
 import com.bytechef.component.definition.ComponentDefinition;
-import com.bytechef.platform.component.registry.definition.ComponentDefinitionWrapper;
-import com.bytechef.platform.component.registry.definition.ComponentHandlerWrapper;
+import com.bytechef.platform.component.definition.ComponentDefinitionWrapper;
+import com.bytechef.platform.component.definition.ComponentHandlerWrapper;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.ServiceLoader;
